application/requirements.txt
streamlit
requests
fastapi
uvicorn
boto3
python-dotenv
pymupdf